pyproject.toml
[project]
name = "advances-ai-agent"
version = "0.1.0"
description = "Research agent using Groq and Tavily"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "langchain>=0.3.0",
    "langchain-groq>=0.2.0",
    "langgraph>=0.2.0",
    "python-dotenv>=1.0.0",
    "tavily-python>=0.5.0",
]